Move version switcher to language bar
